Added archived pre-workshop survey data
Survey data archived as of 2016-11-04 was added to the repo. All survey answers were converted to numerical values. See the key in this repo for the corresponding answer choices/numerical values. Unique identifiers have been removed. Column headings have been added. See the key in this repo to match column headings to the survey questions they correspond to.